Apply Panache accessor generation for public fields to all Hibernate entities, even non-Panache ones
