+++ extensions/panache/hibernate-reactive-panache/deployment/src/main/java/io/quarkus/hibernate/reactive/panache/deployment/PanacheHibernateResourceProcessor.java
@@ -1,6 +1,5 @@
 package io.quarkus.hibernate.reactive.panache.deployment;
 
-import java.lang.reflect.Modifier;
 import java.util.Collections;
 import java.util.HashSet;
 import java.util.List;
@@ -14,12 +13,9 @@ import org.hibernate.reactive.mutiny.Mutiny;
 import org.jboss.jandex.AnnotationInstance;
 import org.jboss.jandex.ClassInfo;
 import org.jboss.jandex.DotName;
-import org.jboss.jandex.FieldInfo;
 
 import io.quarkus.arc.deployment.UnremovableBeanBuildItem;
 import io.quarkus.arc.deployment.ValidationPhaseBuildItem;
-import io.quarkus.bootstrap.classloading.ClassPathElement;
-import io.quarkus.bootstrap.classloading.QuarkusClassLoader;
 import io.quarkus.builder.BuildException;
 import io.quarkus.deployment.Feature;
 import io.quarkus.deployment.annotations.BuildProducer;
@@ -30,19 +26,12 @@ import io.quarkus.deployment.builditem.BytecodeTransformerBuildItem;
 import io.quarkus.deployment.builditem.CombinedIndexBuildItem;
 import io.quarkus.deployment.builditem.FeatureBuildItem;
 import io.quarkus.deployment.util.JandexUtil;
-import io.quarkus.gizmo.DescriptorUtils;
 import io.quarkus.hibernate.orm.deployment.AdditionalJpaModelBuildItem;
-import io.quarkus.hibernate.orm.deployment.HibernateEnhancersRegisteredBuildItem;
 import io.quarkus.hibernate.reactive.panache.PanacheEntity;
 import io.quarkus.hibernate.reactive.panache.PanacheEntityBase;
 import io.quarkus.hibernate.reactive.panache.PanacheRepository;
 import io.quarkus.hibernate.reactive.panache.PanacheRepositoryBase;
-import io.quarkus.panache.common.deployment.EntityField;
-import io.quarkus.panache.common.deployment.EntityModel;
-import io.quarkus.panache.common.deployment.MetamodelInfo;
-import io.quarkus.panache.common.deployment.PanacheEntityClassesBuildItem;
-import io.quarkus.panache.common.deployment.PanacheFieldAccessEnhancer;
-import io.quarkus.panache.common.deployment.PanacheJpaEntityAccessorsEnhancer;
+import io.quarkus.panache.common.deployment.HibernateEnhancersRegisteredBuildItem;
 import io.quarkus.panache.common.deployment.PanacheJpaEntityOperationsEnhancer;
 import io.quarkus.panache.common.deployment.PanacheMethodCustomizer;
 import io.quarkus.panache.common.deployment.PanacheMethodCustomizerBuildItem;
@@ -131,61 +120,6 @@ public final class PanacheHibernateResourceProcessor {
             String entityClassName = entityClass.get().name().toString();
             transformers.produce(new BytecodeTransformerBuildItem(true, entityClassName, entityOperationsEnhancer));
         }
-    }
-
-    @BuildStep
-    @Consume(HibernateEnhancersRegisteredBuildItem.class)
-    void replaceFieldAccesses(CombinedIndexBuildItem index,
-            BuildProducer<BytecodeTransformerBuildItem> transformers,
-            BuildProducer<PanacheEntityClassesBuildItem> fieldAccessEnhancedEntityClasses,
-            List<PanacheEntityClassBuildItem> entityClasses) {
-        MetamodelInfo modelInfo = new MetamodelInfo();
-
-        // Generate accessors for public fields in entities
-        PanacheJpaEntityAccessorsEnhancer entityAccessorsEnhancer = new PanacheJpaEntityAccessorsEnhancer(index.getIndex(),
-                modelInfo);
-        for (PanacheEntityClassBuildItem entityClass : entityClasses) {
-            String entityClassName = entityClass.get().name().toString();
-            modelInfo.addEntityModel(createEntityModel(entityClass.get()));
-            transformers.produce(new BytecodeTransformerBuildItem(true, entityClassName, entityAccessorsEnhancer));
-        }
-
-        if (modelInfo.hasEntities()) {
-            // Share with other extensions that we generated accessors for some classes
-            fieldAccessEnhancedEntityClasses.produce(new PanacheEntityClassesBuildItem(modelInfo.getEntityClassNames()));
-
-            // Replace field access in application code with calls to accessors
-            Set<String> entityClassNamesInternal = new HashSet<>();
-            for (String entityClassName : modelInfo.getEntityClassNames()) {
-                entityClassNamesInternal.add(entityClassName.replace(".", "/"));
-            }
-
-            PanacheFieldAccessEnhancer panacheFieldAccessEnhancer = new PanacheFieldAccessEnhancer(modelInfo);
-            QuarkusClassLoader tccl = (QuarkusClassLoader) Thread.currentThread().getContextClassLoader();
-            List<ClassPathElement> archives = tccl.getElementsWithResource(META_INF_PANACHE_ARCHIVE_MARKER);
-            for (ClassPathElement i : archives) {
-                for (String res : i.getProvidedResources()) {
-                    if (res.endsWith(".class")) {
-                        String cn = res.replace("/", ".").substring(0, res.length() - 6);
-                        transformers.produce(
-                                new BytecodeTransformerBuildItem(cn, panacheFieldAccessEnhancer, entityClassNamesInternal));
-                    }
-                }
-            }
-        }
-    }
-
-    private EntityModel createEntityModel(ClassInfo classInfo) {
-        EntityModel entityModel = new EntityModel(classInfo);
-        for (FieldInfo fieldInfo : classInfo.fields()) {
-            String name = fieldInfo.name();
-            if (Modifier.isPublic(fieldInfo.flags())
-                    && !Modifier.isStatic(fieldInfo.flags())
-                    && !fieldInfo.hasAnnotation(DOTNAME_TRANSIENT)) {
-                entityModel.addField(new EntityField(name, DescriptorUtils.typeToString(fieldInfo.type())));
-            }
-        }
-        return entityModel;
     }
 
     @BuildStep
