+++ extensions/panache/panache-hibernate-common/deployment/src/main/java/io/quarkus/panache/common/deployment/HibernateModelClassCandidatesForFieldAccessBuildItem.java
@@ -1,0 +1,17 @@
+package io.quarkus.panache.common.deployment;
+
+import java.util.Set;
+
+import io.quarkus.builder.item.SimpleBuildItem;
+
+public final class HibernateModelClassCandidatesForFieldAccessBuildItem extends SimpleBuildItem {
+    private final Set<String> allModelClassNames;
+
+    public HibernateModelClassCandidatesForFieldAccessBuildItem(Set<String> allModelClassNames) {
+        this.allModelClassNames = allModelClassNames;
+    }
+
+    public Set<String> getAllModelClassNames() {
+        return allModelClassNames;
+    }
+}
